refact: create register form component
