+++ if-food-front/src/pages/register/components/form/schema.ts
